requirements.txt
annotated-types==0.7.0
attrs==25.1.0
babel==2.17.0
bibtexparser==2.0.0b8
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clldutils==3.24.1
cloudpathlib==0.20.0
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
csvw==3.5.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
dlinfo==2.0.0
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
filelock==3.17.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kokoro==0.7.9
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
loguru==0.7.3
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
misaki==0.7.9
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
num2words==0.5.14
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
phonemizer-fork==3.3.2
pillow==11.1.0
preshed==3.0.9
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.1
pyperclip==1.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rdflib==7.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scipy==1.15.1
segments==2.2.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soundfile==0.13.1
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.13.1
tabulate==0.9.0
thinc==8.3.4
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
triton==3.2.0
typer==0.15.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
